Ensure sound doesn't run during SSR
