+++ src/components/Game.tsx
@@ -1,5 +1,4 @@
 import { Stage } from '@pixi/react';
-import { sound } from '@pixi/sound';
 import { useEffect, useRef, useState } from 'react';
 import { PixiStaticMap } from './PixiStaticMap';
 import { ConvexProvider, useConvex, useMutation, useQuery } from 'convex/react';
@@ -9,25 +8,9 @@ import { HEARTBEAT_PERIOD } from '../../convex/config';
 import { Id } from '../../convex/_generated/dataModel';
 import dynamic from 'next/dynamic';
 
-// Disabling SSR for PixiViewport, as its dependency tries to access `window`.
+// Disabling SSR for these since they don't work server side.
 const PixiViewport = dynamic(() => import('./PixiViewport'), { ssr: false });
-
-// Some hacky sound code for fun
-sound.add('background',  '/assets/background.mp3');
-let isPlaying = false;
-
-  const keyDownEvent = (event: React.KeyboardEvent<HTMLDivElement>) => {
-    if (event.code === "KeyM") {
-      if (!isPlaying ) {
-        sound.play('background');
-        isPlaying = true;
-      } else {
-        sound.stop('background');
-        isPlaying = false;
-      }
-    }
-  };
-// Some hacky sound code for fun
+const Sound = dynamic(() => import('./Sound'), { ssr: false });
 
 export const Game = ({
   setSelectedPlayer,
@@ -41,33 +24,34 @@ export const Game = ({
   const convex = useConvex();
   const worldState = useQuery(api.players.getWorld, {});
 
-
   const offset = useServerTimeOffset(worldState?.world._id);
   if (!worldState) return null;
   const { players } = worldState;
   return (
-    <div className="container" onKeyDown={keyDownEvent} tabIndex={0}>
-    <Stage width={width} height={height} options={{ backgroundColor: 0x7ab5ff }}>
-      <PixiViewport
-        screenWidth={width}
-        screenHeight={height}
-        worldWidth={worldState.map.tileSetDim}
-        worldHeight={worldState.map.tileSetDim}
-      >
-        <PixiStaticMap map={worldState.map}></PixiStaticMap>
-        <ConvexProvider client={convex}>
-          {players.map((player) => (
-            <Player
-              key={player._id}
-              player={player}
-              offset={offset}
-              tileDim={worldState.map.tileDim}
-              onClick={setSelectedPlayer}
-            />
-          ))}
-        </ConvexProvider>
-      </PixiViewport>
-    </Stage>
+    <div className="container">
+      <Sound>
+        <Stage width={width} height={height} options={{ backgroundColor: 0x7ab5ff }}>
+          <PixiViewport
+            screenWidth={width}
+            screenHeight={height}
+            worldWidth={worldState.map.tileSetDim}
+            worldHeight={worldState.map.tileSetDim}
+          >
+            <PixiStaticMap map={worldState.map}></PixiStaticMap>
+            <ConvexProvider client={convex}>
+              {players.map((player) => (
+                <Player
+                  key={player._id}
+                  player={player}
+                  offset={offset}
+                  tileDim={worldState.map.tileDim}
+                  onClick={setSelectedPlayer}
+                />
+              ))}
+            </ConvexProvider>
+          </PixiViewport>
+        </Stage>
+      </Sound>
     </div>
   );
 };
